Add roles

+++ internal/handlers/callbacks.go
@@ -5,7 +5,7 @@ import (
 	"strings"
 
 	tmsbot "github.com/NikitaDmitryuk/telegram-media-server/internal/bot"
-	tmsdb "github.com/NikitaDmitryuk/telegram-media-server/internal/database"
+	"github.com/NikitaDmitryuk/telegram-media-server/internal/database"
 	"github.com/NikitaDmitryuk/telegram-media-server/internal/lang"
 	tgbotapi "github.com/go-telegram-bot-api/telegram-bot-api/v5"
 	"github.com/sirupsen/logrus"
@@ -15,19 +15,32 @@ func HandleCallbackQuery(bot *tmsbot.Bot, update tgbotapi.Update) {
 	callbackData := update.CallbackQuery.Data
 	chatID := update.CallbackQuery.Message.Chat.ID
 
+	allowed, role := AuthMiddleware(bot, update)
+	if !allowed {
+		bot.SendErrorMessage(chatID, lang.GetMessage(lang.AccessDeniedMsgID))
+		return
+	}
+
 	switch {
 	case strings.HasPrefix(callbackData, "delete_movie:"):
-		movieID := strings.TrimPrefix(callbackData, "delete_movie:")
-		DeleteMovieByID(bot, chatID, movieID)
-	case callbackData == "delete_movie_menu":
-		movies, err := tmsdb.GlobalDB.GetMovieList(context.Background())
-		if err != nil {
-			bot.SendErrorMessage(chatID, lang.GetMessage(lang.GetMovieListErrorMsgID))
-			return
+		if role == database.AdminRole || role == database.RegularRole {
+			movieID := strings.TrimPrefix(callbackData, "delete_movie:")
+			DeleteMovieByID(bot, chatID, movieID)
+		} else {
+			bot.SendErrorMessage(chatID, lang.GetMessage(lang.AccessDeniedMsgID))
 		}
-		SendDeleteMovieMenu(bot, chatID, movies)
+	case callbackData == "delete_movie_menu":
+		if role == database.AdminRole || role == database.RegularRole {
+			movies, err := database.GlobalDB.GetMovieList(context.Background())
+			if err != nil {
+				bot.SendErrorMessage(chatID, lang.GetMessage(lang.GetMovieListErrorMsgID))
+				return
+			}
+			SendDeleteMovieMenu(bot, chatID, movies)
+		} else {
+			bot.SendErrorMessage(chatID, lang.GetMessage(lang.AccessDeniedMsgID))
+		}
 	case callbackData == "list_movies":
-		logrus.Info("Listing movies")
 		ListMoviesHandler(bot, update)
 	default:
 		logrus.Warnf("Unknown callback data: %s", callbackData)
